app/indexer/requirements.txt
telethon==1.36.0
psycopg2-binary==2.9.9
meilisearch==0.31.5
python-dotenv==1.0.1
